Extract virt notification websocket into a hook
In order to share the code of the virtualization actions notifications
with other pages like the pools or networks lists, move the websocket
handling code into a custom useVirtNotification hook.
